requirements.txt
langchain-community
langchain-huggingface
langgraph
python-dotenv
langgraph-checkpoint-sqlite
faiss-cpu
sentence-transformers